latest from coda-oss

+++ externals/coda-oss/modules/c++/math/unittests/test_n_choose_k.cpp
@@ -27,10 +27,10 @@ namespace
 {
 TEST_CASE(testNChooseK)
 {
-    TEST_ASSERT_EQ(math::nChooseK(0, 0), 1);
-    TEST_ASSERT_EQ(math::nChooseK(1, 1), 1);
-    TEST_ASSERT_EQ(math::nChooseK(3, 2), 3);
-    TEST_ASSERT_EQ(math::nChooseK(10, 3), 120);
+  TEST_ASSERT_EQ(math::nChooseK(0, 0), static_cast<size_t>(1));
+  TEST_ASSERT_EQ(math::nChooseK(1, 1), static_cast<size_t>(1));
+  TEST_ASSERT_EQ(math::nChooseK(3, 2), static_cast<size_t>(3));
+  TEST_ASSERT_EQ(math::nChooseK(10, 3), static_cast<size_t>(120));
 }
 
 TEST_CASE(testNLessThanK)
